LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/intercept-multiple-times-expected.txt


PASS navigation.navigate() returns a finished promise that awaits all promises if event.intercept() is called multiple times
